+++ src/main/java/com/beerhouse/entities/Beer.java
@@ -16,7 +16,7 @@ public class Beer implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private String ingredients;
 	private String alcoholContent;
@@ -27,7 +27,7 @@ public class Beer implements Serializable {
 		
 	}
 
-	public Beer(Integer id, String name, String ingredients, String alcoholContent, BigDecimal price, String category) {
+	public Beer(Long id, String name, String ingredients, String alcoholContent, BigDecimal price, String category) {
 		this.id = id;
 		this.name = name;
 		this.ingredients = ingredients;
@@ -36,11 +36,11 @@ public class Beer implements Serializable {
 		this.category = category;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
