+++ src/main/java/com/beerhouse/repositories/BeerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.beerhouse.entities.Beer;
 
 @Repository
-public interface BeerRepository extends JpaRepository<Beer, Integer> {
+public interface BeerRepository extends JpaRepository<Beer, Long> {
 
 }
